feat: build responsive header - #2
header contains the branding, navbar with links that routes user to the 3 current pages, and a shopping cart button that would take the user to the shopping cart page. this fixes #2. testing of routing will be done at the app level (will be done in #14). #7 will implement the count of current items inside the cart, which will also impact the header