feat: add json file containing in-stock products
this mocks what returns after we fetch and parse from the backend that we will set up. i chose not to use a json server because this then needs to be deployed. might as well just wait until i learn how to set up an express js relational db and then refactor so that this site can actually be deployed (client-side only) on netlify or something. out of stock products do not have photos so no point including. if i wanted to display difference in UI, i can set one of these products as 'false' for isInStock