feat: build individual products page for mvp.
we have not added a scroll area on wider screens for the product images for mvp due to difficulty in implementing - this will be fixed in #24. however, on narrower screens, we use a single column layout for the top section, with carousel of images. leveraged mantine's carousel component. ensure responsiveness with spot testing using various sizes and both landscape and portrait orientations in chrome devtools. fixes #6