feat: build empty shopping cart page.
the empty shopping cart displaying some copy and a routing link back to products page is a separate subcomponent of the CartPage component (called EmptyCart). this is so that it is displayed if the items array prop passed into CartPage parent component is empty. we would otherwise display all of the items. this is one half of issue #8.