refactor: use <Outlet /> in an App.jsx entry point - fixes #23
this prevents having to import and re-render the Header and Footer on every page (keeping codebase DRY). this also enables scalability for future pages (like the About Us and Sustainability pages on the current Shopify site). this will also open the door for using a global state from a parent component (we couldn't really do this in main.jsx because there isn't a component to wrap around a useState hook), enabling #7, and an actual, functional #8 non-empty cart