feat: add simple modal that displays when user adds item to cart. fixes #33.
not much config for mvp, so does not align with the design of current site. #36 will be done to properly place the modal. it might be that we sack using mantine and build the modal from scratch to have better control of its positioning with css. this mvp version meets functional requirements, including being responsive (no media queries required as it is in the centre of the page and is quite narrow - tested on browser.