feat: build workflow for adding item to cart (#38)
* feat: add counter to the top right of the shopping cart icon in the header. adding and removing items updates the counter through calculation using the state of items in cart (global), passed as prop from App.jsx. fixes #34.

* feat: add simple modal that displays when user adds item to cart. fixes #33.

not much config for mvp, so does not align with the design of current site. #36 will be done to properly place the modal. it might be that we sack using mantine and build the modal from scratch to have better control of its positioning with css. this mvp version meets functional requirements, including being responsive (no media queries required as it is in the centre of the page and is quite narrow - tested on browser.

* feat: the 'add to cart' button, on click, will say 'adding' until it is added to the cart. then, when it is, it says 'added to cart' for 3 seconds before reverting to default state ('add to cart'). fixes #32.